mvoe state merging to separate file
